requirements.txt
streamlit
python-dotenv
google-generativeai
nba_api
boto3
psycopg2
grpcio==1.49.0